pyproject.toml
[project]
name = "guarddo-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Daniil",email = "dshpeka@24hsoft.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram (>=3.22.0,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
